backend/app/platforms/deriv.py

from playwright.sync_api import sync_playwright, TimeoutError as PWTimeout
import os, time, logging

logger = logging.getLogger("deriv" + "_fetcher")

LOGIN_URL = "https://deriv.com/en/login"
TRADE_URL = "https://deriv.com/trading"

class DerivFetcher:
    def __init__(self, headless=True, proxy=None):
        self.headless = headless
        self.proxy = proxy
        self.playwright = None
        self.browser = None
        self.context = None
        self.page = None

    def start(self):
        self.playwright = sync_playwright().start()
        launch_args = ["--no-sandbox"]
        self.browser = self.playwright.chromium.launch(headless=self.headless, args=launch_args)
        self.context = self.browser.new_context()
        self.page = self.context.new_page()
        self.page.set_viewport_size({"width": 1366, "height": 768})

    def stop(self):
        try:
            if self.context: self.context.close()
            if self.browser: self.browser.close()
        finally:
            if self.playwright: self.playwright.stop()

    def login(self, username: str = None, password: str = None, timeout=15000):
        """Open the login page and sign in. Update selectors for the platform."""
        if not self.page:
            self.start()
        try:
            self.page.goto(LOGIN_URL, timeout=timeout)
            # TODO: replace the selectors below with platform-specific selectors using Playwright inspector
            # Example placeholder selectors:
            # self.page.fill("input[name='email']", username)
            # self.page.fill("input[name='password']", password)
            # self.page.click("button[type='submit']")
            time.sleep(2)
            return True
        except PWTimeout:
            logger.exception("Login timeout")
            return False

    def fetch_price(self, pair: str, timeout=10000):
        """Navigate to the trading page for the pair and scrape the current price.
        Must update CSS/XPath selectors for the live site.
        Returns dict: {'pair': pair, 'price': float, 'ts': int}
        """
        if not self.page:
            self.start()
        try:
            # Open trade page (may need pair-specific URL)
            self.page.goto(TRADE_URL, timeout=timeout)
            # Wait for price selector - UPDATE THIS SELECTOR
            # Examples of selectors you must replace using the Playwright inspector:
            # price_text = self.page.inner_text(".price-selector")
            # Or use xpath: price_text = self.page.inner_text("xpath=//div[@class='price']")
            self.page.wait_for_timeout(500)  # placeholder - wait for content
            # Placeholder naive extraction - replace with specific selector logic in production
            text = self.page.inner_text('body')[:200]
            import re, time
            m = re.search(r"(\d+\.?\d*)", text)
            if m:
                price = float(m.group(1))
            else:
                price = 0.0
            return {'pair': pair, 'price': price, 'ts': int(time.time())}
        except Exception as e:
            logger.exception('fetch_price error: %s', e)
            return {'pair': pair, 'price': None, 'ts': int(time.time())}
