backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
cryptography
